Add PositionChecker section to README
Added a new section titled "PositionChecker" to the README.md file. This section includes a brief description of the "PositionChecker" web app, which checks and visualizes the validity of X and Y values with a smooth background color transition. Also added a link to the "PositionChecker" web app (PositionChecker/index.html).